==> 0-CreateDataModel.ipynb <==
# Create Data Model

This notebook is only to prepare the Neo4j database for uploading the data:
- it deletes all existing nodes and labels (reset of the database)
- it create indexes and unique id constraints for both Duns and Erp Labels
- Connection to Neo4j is a prerequisite

### Connect to graph

import py2neo
graph = py2neo.Graph(user='neo4j',password='admin')

### Detach Delete

import py2neo
graph = py2neo.Graph(user='neo4j',password='admin')
graph.run('''
Match (n)
Detach Delete (n);
''')

### Create Unique Index on DunsId

graph.run('CREATE CONSTRAINT ON (c:DunsID) ASSERT c.dunsid IS UNIQUE;')
graph.run('CREATE CONSTRAINT ON (c:ErpID) ASSERT c.companysystemid IS UNIQUE;')


### Create Index on Company Name

graph.run('CREATE INDEX ON :DunsID(companyname);')
graph.run('CREATE INDEX ON :ErpID(companyname);')
graph.run('CREATE INDEX ON :ErpID(dunsid);')
