==> 00-Run notebooks 0 to 4.ipynb <==
## Click Cell / Run All

Run upload queries on duns

%run 02a-LoadDunsData.ipynb

Run upload query on erp

%run 02b-LoadErpData.ipynb

Run update query to calculate spend, etc...

%run 03-Run Updates.ipynb
