==> 01b - CreateCypherUploadQueries.ipynb <==
# Data quality for linkurious

sourcepath = 'C:/_PYDAI/1-Data/'
neo4jimportpath='C:/_PYDAI/2-Neo4j_db/duns.graphdb/import/'
os.chdir(sourcepath)
outpath='C:/_PYDAI/8-PythonProjects/uploadinneo4j/'

filename__duns_node='DunsID_normalized.csv'
filename_duns_edge='DunsRelations_normalized.csv'
filename__erp_node='ErpID_normalized.csv'
filename_erp_edge='ErpRelations_normalized.csv'

script_duns_node='upload_duns_node.cql'
script_duns_edge='upload_duns_edge.cql'
script_erp_node='upload_erp_node.cql'
script_erp_edge='upload_erp_edge.cql'

# Write upload neo4j queries

# For nodes

### Function

def createlineuploadnode(colname,nodename,neo4jname=None):
    colname =str(colname).lower().lstrip().rstrip().replace(' ','')
    if neo4jname is None:
        neo4jname = colname
    s="{0}.{1} = {2}.{3}".format(nodename,neo4jname,'csvLine',neo4jname)
    return s
def createfullquery(filename,path=neo4jimportpath,labelname='DunsID',nodeid='duns'):
    s='''
    USING PERIODIC COMMIT 10000
    LOAD CSV WITH HEADERS FROM 'file:///{0}' AS csvLine
    MERGE (n:{labelname}{{ {nodeid}: csvLine.{nodeid} }})
    SET
    '''
    s=s.format(filename,labelname=labelname,nodeid=nodeid)
    mycols=pd.read_csv(path+filename,nrows=5).columns
    for c in mycols:
        r = createlineuploadnode(c,nodename='n')
        r="\t"+r+",\n"
        s+=r
    s=s.rstrip(',\n') +';'
    return s

### Duns

query = createfullquery(filename=filename__duns_node,path=neo4jimportpath,labelname='DunsID',nodeid='duns')
print(query)
myfile=open(outpath+script_duns_node,'w').close()
print(query,file=open(outpath+script_duns_node,'a'))

### ERP

query = createfullquery(filename=filename__erp_node,path=neo4jimportpath,labelname='ErpID',nodeid='erpid')
print(query)
myfile=open(outpath+script_erp_node,'w').close()
print(query,file=open(outpath+script_erp_node,'a'))

## Relations

s='''
MATCH (a:DunsID),(b:DunsID)
WHERE b.dunsid = a.parentdunsid and not a.parentdunsid=a.dunsid
MERGE (a)-[r:isChildOf]->(b)
WITH a,r
SET r.source_relation = a.source_relation
'''
print(s)

def createedgequery(filename,path=neo4jimportpath,filerelation='isChildOf',sourcelabel='DunsID',targetlabel='DunsID',
                    neo4jsourceid='duns',neo4jtargetid='duns'):
    df=pd.read_csv(path+filename,nrows=5)
    mycols=df.columns
    filesourceid=mycols[0]
    filetargetid=mycols[2]
    p='''USING PERIODIC COMMIT 10000
    LOAD CSV WITH HEADERS FROM 'file:///{filename}' AS csvLine
    MERGE (a:{sourcelabel}{{ {neo4jsourceid}: csvLine.{filesourceid} }})-[r:csvLine.{filerelation}]->(b:{targetlabel}{{ {neo4jtargetid}: csvLine.{filetargetid} }})
    WITH r
    SET
    '''
    p=p.format(filename=filename,
        sourcelabel=sourcelabel,neo4jsourceid=neo4jsourceid,
             filesourceid=filesourceid,filerelation='filerelation',
             targetlabel=targetlabel,neo4jtargetid=neo4jtargetid,filetargetid=filetargetid)
    for c in mycols:
        r = createlineuploadnode(c,nodename='r')
        r="\t"+r+",\n"
        p+=r
    p=p.rstrip(',\n') +';'
    return p

### Duns

query = createedgequery(filename=filename_duns_edge,
                        path=neo4jimportpath,
                       sourcelabel='DunsID',
                       neo4jsourceid='duns',
                        filerelation='isChildOf',
                       targetlabel='DunsID',
                       neo4jtargetid='duns')
print(query)
myfile=open(outpath+script_duns_edge,'w').close()
print(query,file=open(outpath+script_duns_edge,'a'))

### Erp

query = createedgequery(filename=filename_erp_edge,
                        path=neo4jimportpath,
                       sourcelabel='ErpID',
                       neo4jsourceid='erpid',
                        filerelation='isChildOf',
                       targetlabel='DunsID',
                       neo4jtargetid='duns')
print(query)
myfile=open(outpath+script_erp_edge,'w').close()
print(query,file=open(outpath+script_erp_edge,'a'))
