==> 02a-LoadDunsData.ipynb <==
# Load Duns Data

Find and format File

filepath='../../1-Data/'
filename='dunswithlatlng.csv'
dunsdata=pd.read_csv(filepath+filename,sep='|',encoding='utf-8',
                    dtype={
                        'companyname':str,
                          'countrycode':str,
                          'countryname':str,
                           'street':str,
                           'city':str,
                          'postalcode':str,
                          'registerid':str,
                          'source_level':str,
                            'source_relation':str,
                          'isglobalroot':bool,
                            'state':str,
                        'lat':str,
                        'lng':str
                    })

dunsdata=pd.read_excel('testexcel.xlsx')

dunsdata.shape

dunsdata.dropna(subset=['parentdunsid'],inplace=True)
dunsdata=dunsdata.loc[dunsdata['parentdunsid'].isin(dunsdata['dunsid'])]
dunsdata.fillna('notavailable',inplace=True)

import neatmartinet as nm
for c in ['dunsid','parentdunsid']:
    dunsdata[c]=dunsdata[c].apply(lambda r:nm.convert_int_to_str(n=r,fillwithzeroes=9))
print(nm.checkuniqueid(dunsdata['dunsid']))

neopath='../../Neo4j_db/duns.graphdb/import/'
dunsdata.to_csv(neopath+'dunsdata_commasep.csv',encoding='utf-8',sep=',',index=False)

# Load in neo4j

### Connect to graph

import py2neo
graph = py2neo.Graph(user='neo4j',password='admin')

#### Load Duns

query_uploadduns='''
USING PERIODIC COMMIT 10000
  LOAD CSV WITH HEADERS FROM 'file:///dunsdata_commasep.csv' AS row
    MERGE (c:DunsID{dunsid: row.dunsid})
    ON CREATE SET
        c.dunsid=row.dunsid,
        c.companyname=row.companyname,
        c.countrycode=row.countrycode,
        c.countryname=row.countryname,
        c.street=row.street,
        c.city=row.city,
        c.postalcode=row.postalcode,
        c.registerid=row.registerid,
        c.source_level=row.source_level,
        c.parentdunsid=row.parentdunsid,
        c.source_relation=row.source_relation,
        c.isglobalroot=row.isglobalroot,
        c.lat=row.lat,
        c.lng = row.lng,
        c.creationdate=timestamp(),
        c.lastseen=timestamp()
    ON MATCH SET
        c.lastseen=timestamp(),
        c.lat=row.lat,
        c.lng = row.lng;
'''
a=graph.run(query_uploadduns)
print(a.stats())

#### Create Relations

query_dunsischildof='''
MATCH (a:DunsID),(b:DunsID)
WHERE b.dunsid = a.parentdunsid and not a.parentdunsid=a.dunsid
MERGE (a)-[r:isChildOf]->(b)
WITH a,r
SET r.source_relation = a.source_relation
'''
a=graph.run(query_dunsischildof)
print(a.stats())

# TO DO : 
- Write update queries for country root
- Wrie RegisterId relation

# Close
