==> 03-Run Updates.ipynb <==
# Run Updates

### Connect to Graph

import py2neo
graph = py2neo.Graph(user='neo4j',password='admin')

# Global Root Update

# Count the number of relationships
query_attributenfathers='''
MATCH (n:DunsID)
OPTIONAL MATCH (n)-[r:isChildOf]->(c)
WITH n, count(c) AS nfathers
SET n.nfathers = nfathers
'''
a=graph.run(query_attributenfathers)
print(a.stats())
#if there are no relationships then update the isglobalroot attribute
query_attributeglobalroot='''
MATCH (n:DunsID)
WITH n, (CASE n.nfathers WHEN 0 THEN True Else False END) as root
SET n.isglobalroot = root
'''
a=graph.run(query_attributeglobalroot)
print(a.stats())

#Update the labels accordingly
query_labelglobalroot1='''
MATCH (n:DunsID{isglobalroot:False})
REMOVE n:GlobalRoot, n.rootname'''
a=graph.run(query_labelglobalroot1)
print(a.stats())

query_labelglobalroot2='''
MATCH (m:DunsID{isglobalroot:True})
Set m:GlobalRoot,m.rootname = m.companyname;
'''
a=graph.run(query_labelglobalroot2)
print(a.stats())

### Global Root and Direct Parent attribute update for export

#### Duns5

query_duns5='''
MATCH (n:DunsID)-[:isChildOf*0..]->(r:GlobalRoot)
SET n.duns5_calc=r.dunsid;
'''
a=graph.run(query_duns5)
print(a.stats())

#### Parent Duns

query_duns3='''
MATCH (n:DunsID{isglobalroot:False})-[:isChildOf]->(r:DunsID)
SET n.parentdunsid=r.dunsid;
'''
a=graph.run(query_duns3)
print(a.stats())
query_duns3_bis='''
MATCH (n:GlobalRoot)
SET n.parentdunsid=n.dunsid;'''
a=graph.run(query_duns3_bis)
print(a.stats())

# Spent

### Per Duns Node

query_nodeturnover='''
MATCH(m:DunsID)
SET m.nodeturnover=0
with count(m) as temp
MATCH (n:DunsID)<-[:hasDuns]-(s:ErpID)
WITH n,SUM(s.turnover) as erpturnover
SET n.nodeturnover=erpturnover;
'''
a = graph.run(query_nodeturnover)
print(a.stats())

query_nodeordervolume='''
MATCH(m:DunsID)
SET m.nodeordervolume=0
with count(m) as temp
MATCH (n:DunsID)<-[:hasDuns]-(s:ErpID)
WITH n,SUM(s.ordervolume) as erpordervolume
SET n.nodeordervolume=erpordervolume;
'''
a = graph.run(query_nodeordervolume)
print(a.stats())

### Aggregate turnover

old_query='''
MATCH (n:DunsID)<-[:isChildOf*0..]-(child:DunsID)
WITH n,sum(child.nodeturnover) AS value
SET n.aggregateturnover = value
'''

query_leaves='''
MATCH path=(r:GlobalRoot)<-[:isChildOf*0..]-(b:DunsID), (b)<-[:isChildOf]-(l:DunsID)
WITH b, length(path) as depth, collect(l) as ls
ORDER BY depth desc
SET b.aggregateturnover = REDUCE(acc = COALESCE(b.nodeturnover,0), l IN ls | acc + COALESCE(l.aggregateturnover, l.nodeturnover, 0))
'''
a=graph.run(query_leaves)
print(a.stats())

query_root='''
MATCH (n:GlobalRoot)
OPTIONAL MATCH (n)<-[:isChildOf]-(child:DunsID)
WITH n,sum(child.aggregateturnover) AS value
SET n.aggregateturnover = value+n.nodeturnover
'''
a=graph.run(query_root)
print(a.stats())

### Aggregate ordervolume

query_leaves='''
MATCH path=(r:GlobalRoot)<-[:isChildOf*0..]-(b:DunsID), (b)<-[:isChildOf]-(l:DunsID)
WITH b, length(path) as depth, collect(l) as ls
ORDER BY depth desc
SET b.aggregateordervolume = REDUCE(acc = COALESCE(b.nodeordervolume,0), l IN ls | acc + COALESCE(l.aggregateordervolume, l.nodeordervolume, 0))
'''
a=graph.run(query_leaves)
print(a.stats())

query_root='''
MATCH (n:GlobalRoot)
OPTIONAL MATCH (n)<-[:isChildOf]-(child:DunsID)
WITH n,sum(child.aggregateordervolume) AS value
SET n.aggregateordervolume = value+n.nodeordervolume
'''
a=graph.run(query_root)
print(a.stats())
