==> 3-LoadErpData.ipynb <==
# Load Erp Data

Find and format File

erpdata=pd.read_csv('../1-AD/3-DUNS/duns_denormalized.csv',sep='|',encoding='utf-8',
                    dtype={
                        'companyname':str,
                          'countrycode':str,
                          'countryname':str,
                           'street':str,
                           'city':str,
                          'postalcode':str,
                          'registerid':str,
                          'source_level':str,
                            'source_relation':str,
                          'isglobalroot':bool})

filepath='../1-AD/6-sourcingreport/'
filename='2017-10-18 ERP supplier Linkurious_2POR.xlsx'
erpdata=pd.read_excel(filepath+filename)

erpdata.dropna(subset=['companysystemid'],inplace=True)
erpdata.fillna('notavailable',inplace=True)

import neatmartinet as nm
for c in ['dunsid']:
    erpdata[c]=erpdata[c].apply(lambda r:nm.convert_int_to_str(n=r,fillwithzeroes=9))

neoimportpath='../../Neo4j/sourcingreport.graphdb/import/'
erpdata.to_csv(neoimportpath+'erpdata_commasep.csv',encoding='utf-8',sep=',',index=False)

erpdata['postalcode']='notavailable'
erpdata['street']='notavailable'
erpdata['registerid']='notavailable'

erpdata.isnull().sum()

# Load in neo4j

### Connect to graph

import py2neo
graph = py2neo.Graph(user='neo4j',password='admin')

#### Load Erp

query_uploaderp='''
USING PERIODIC COMMIT 10000
  LOAD CSV WITH HEADERS FROM 'file:///erpdata_commasep.csv' AS row
    MERGE (c:ErpID{companysystemid: row.companysystemid})
    ON CREATE SET
        c.companysystemid=row.companysystemid,
        c.systemid=row.systemid,
        c.companylocalid=row.companylocalid,
        c.dunsid=row.dunsid,
        c.companyname=row.companyname,
        c.countrycode=row.countrycode,
        c.countryname=row.countryname,
        c.street=row.street,
        c.city=row.city,
        c.postalcode=row.postalcode,
        c.registerid=row.registerid,
        c.turnover=toFloat(row.turnover),
        c.ordervolume=toFloat(row.ordervolume),
        c.creationdate=timestamp(),
        c.lastseen=timestamp()
    ON MATCH SET
        c.lastseen=timestamp(),
        c.spent=toFloat(row.spent);
'''
a=graph.run(query_uploaderp)
print(a.stats())

#### Create Relations

query_erphasduns='''
MATCH (a:ErpID),(b:DunsID)
WHERE b.dunsid = a.dunsid
MERGE (a)-[:hasDuns]->(b)
'''
a=graph.run(query_erphasduns)
print(a.stats())

# TO DO : 
- Create Country relations
- Write relation for RegisterId
- Deal with NDM999999
