==> 04-Xport Data to csv.ipynb <==
import py2neo
graph = py2neo.Graph(user='neo4j',password='admin')

### Export Duns

filename='dunsexport.csv'

query='''
MATCH (n:DunsID)
RETURN properties(n) as data
'''
result=graph.data(query)
x=pd.Series(result).apply(lambda r:r['data'])
duns=x.apply(lambda r:pd.Series(r))
duns.to_csv('../../Neo4j/sourcingreport.graphdb/'+filename,encoding='utf-8',sep='|',index=False)
print('filename:',filename)
print('shape of table',duns.shape)
print('total node spent',duns['nodespent'].sum())
print('node / aggregate spent coherent:',duns['nodespent'].sum()==duns.loc[duns['isglobalroot']==True,'aggregatespent'].sum())

### Export ERP

filename='erpexport.csv'

query='''
MATCH (n:ErpID)
RETURN properties(n) as data
'''
result=graph.data(query)
x=pd.Series(result).apply(lambda r:r['data'])
erp=x.apply(lambda r:pd.Series(r))
erp.to_csv('../../Neo4j/sourcingreport.graphdb/'+filename,encoding='utf-8',sep='|',index=False)
print('filename:',filename)
print('shape of table',erp.shape)
print('total erp spent',erp['spent'].sum())
print('duns / erp aggregate spent coherent:',duns['nodespent'].sum()==erp['spent'].sum())
